updated student details layout, modified writeNewStudent implementation

+++ app/src/main/java/com/example/robertszekely/teacherplanner/models/Student.java
@@ -17,7 +17,7 @@ public class Student {
     private String email;
     private String phoneNumber;
     private String group;
-    private double progress = 0;
+    private int progress = 0;
 
     public Student() {
         //Default constructor required for calls to DataSnapshop.getValue(Student.class)
@@ -71,11 +71,11 @@ public class Student {
         this.group = group;
     }
 
-    public double getProgress() {
+    public int getProgress() {
         return progress;
     }
 
-    public void setProgress(double progress) {
+    public void setProgress(int progress) {
         this.progress = progress;
     }
 }
